+++ src/main/java/subway/station/service/dto/LineFindAllResponse.java
@@ -1,4 +1,4 @@
-package subway.station.web.dto;
+package subway.station.service.dto;
 
 import lombok.Builder;
 import lombok.Getter;
