+++ src/main/java/subway/station/service/dto/StationFindAllResponse.java
@@ -1,4 +1,4 @@
-package subway.station.web.dto;
+package subway.station.service.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,13 +6,13 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class LineUpdateResponse {
+public class StationFindAllResponse {
+    private Long id;
     private String name;
-    private String color;
 
     @Builder
-    public LineUpdateResponse(String name, String color) {
+    public StationFindAllResponse(Long id, String name) {
+        this.id = id;
         this.name = name;
-        this.color = color;
     }
 }
